requirement.txt
asgiref==3.3.1
beautifulsoup4==4.10.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-bootstrap-form==3.4
django-bootstrap4==21.2
django-crispy-forms==1.14.0
django-froala-editor==4.0.9
idna==2.10
phonenumbers==8.12.20
Pillow==9.0.1
pycountry==20.7.3
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.3.1
sqlparse==0.4.1
twilio==6.54.0
tzdata==2021.5
urllib3==1.26.4
